Time: 0 ms (100%), Space: 45.6 MB (89.8%) - LeetHub

+++ 0450-delete-node-in-a-bst/0450-delete-node-in-a-bst.java
@@ -35,8 +35,10 @@ class Solution {
                 return root;
             }
         }
-        root.left = deleteUtil(root.left, key);
-        root.right = deleteUtil(root.right, key);
+        if(key <= root.val)
+            root.left = deleteUtil(root.left, key);
+        if(key >= root.val)
+            root.right = deleteUtil(root.right, key);
         return root;
     }
     public TreeNode deleteNode(TreeNode root, int key) {
